Refactor Redux actions and reducers
